Restructure atoms
