+++ src/components/atom/AppHeader/index.tsx
@@ -1,7 +1,8 @@
 import { CompositeNavigationProp } from "@react-navigation/core";
 import React, { FC, PropsWithChildren } from "react";
-import { StyleSheet, View } from "react-native";
-import { colorsPalette as cp } from "../../constants";
+import { View } from "react-native";
+import styles from "./styles";
+
 interface AppHeaderProps {
   navigation?: CompositeNavigationProp<any, any>;
 }
@@ -10,13 +11,5 @@ const AppHeader: FC<PropsWithChildren<AppHeaderProps>> = ({ children }) => {
   const s = styles();
   return <View style={s.container}>{children}</View>;
 };
-
-const styles = () =>
-  StyleSheet.create({
-    container: {
-      backgroundColor: cp.main,
-      height: 54,
-    },
-  });
 
 export default AppHeader;
