+++ src/components/atom/AppHeader/styles.ts
@@ -1,0 +1,12 @@
+import { StyleSheet } from "react-native";
+import { colorsPalette as cp } from "../../../constants";
+
+const styles = () =>
+  StyleSheet.create({
+    container: {
+      backgroundColor: cp.main,
+      height: 54,
+    },
+  });
+
+export default styles;
