+++ src/components/atom/ButtonHeader/styles.ts
@@ -1,0 +1,13 @@
+import { StyleSheet } from "react-native";
+
+const styles = () =>
+  StyleSheet.create({
+    container: {
+      height: 54,
+      width: 54,
+      justifyContent: "center",
+      alignItems: "center",
+    },
+  });
+
+export default styles;
